comms: rename dir from comms_msat to comms
