+++ src/comms/radio.py
@@ -1,0 +1,9 @@
+class Radio:
+    def __init__(self):
+        print('called radio init!')
+
+    def transmit(self):
+        print('radio is transmitting!')
+
+    def receive(self):
+        print('radio is receiving!')
